Moving PredictNext to the transforms package
